Create engine.ts file and refactor main.ts
